Reparado el archivo de admin, movidas las vistas a admin en la app users
